Fix compatibility issues with monitors
Remove event listeners to detect `monitor_touch` or `monitor_resize`. When using the built-in `monitor` program on the computer those events get fired and requeued as `mouse_click` and `term_resize` respectively, thus causing double firing. Also force the program to check the terminal size and to quit when the screen is too small to use.